+++ platform/platform-impl/src/com/intellij/ide/ui/laf/intellij/WinIntelliJSpinnerUI.java
@@ -176,7 +176,7 @@ public class WinIntelliJSpinnerUI extends DarculaSpinnerUI {
             g2.setColor(UIManager.getColor("Button.intellij.native.borderColor"));
           }
 
-          Path2D border = new Path2D.Double(Path2D.WIND_EVEN_ODD);
+          Path2D border = new Path2D.Float(Path2D.WIND_EVEN_ODD);
           border.append(outerRect, false);
 
           Rectangle innerRect = new Rectangle(outerRect);
